% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-helpers.R
\name{mod_call_satscan}
\alias{mod_call_satscan}
\title{Run Scan}
\usage{
mod_call_satscan(
  .data,
  filename,
  directory,
  sslocation,
  ssbatchfilename,
  satscan_version,
  scan_for,
  latitude,
  longitude,
  gam_based,
  area,
  vals,
  output
)
}
\description{
Run Scan
}
\keyword{internal}
