% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-helpers.R
\name{mod_display_input_variables}
\alias{mod_display_input_variables}
\title{Display input variables dynamically, according to UI for screening}
\usage{
mod_display_input_variables(
  .data,
  analysis_scope = c("single", "multiple"),
  ns
)
}
\description{
Display input variables dynamically, according to UI for screening
}
\keyword{internal}
