% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-spatial-scan.R
\name{module_server_run_spatial_scan}
\alias{module_server_run_spatial_scan}
\title{Module server for data upload}
\usage{
module_server_run_spatial_scan(id, .data)
}
\arguments{
\item{id}{Module ID}
}
\description{
Module server for data upload
}
\keyword{internal}
