% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-upload.R
\name{module_server_upload}
\alias{module_server_upload}
\title{Module server for data upload}
\usage{
module_server_upload(id)
}
\arguments{
\item{id}{Module ID}
}
\description{
Module server for data upload
}
\keyword{internal}
