% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-data-wrangling.R
\name{module_server_wrangle_data}
\alias{module_server_wrangle_data}
\title{Module server for data upload}
\usage{
module_server_wrangle_data(id, data)
}
\arguments{
\item{id}{Module ID}
}
\description{
Module server for data upload
}
\keyword{internal}
