% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-spatial-scan.R
\name{module_ui_run_spatial_scan}
\alias{module_ui_run_spatial_scan}
\title{Module UI for data upload}
\usage{
module_ui_run_spatial_scan(id)
}
\arguments{
\item{id}{Module ID}
}
\description{
Module UI for data upload
}
\keyword{internal}
