% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_upload.R
\name{module_ui_upload}
\alias{module_ui_upload}
\title{Module UI for data upload}
\usage{
module_ui_upload(id)
}
\arguments{
\item{id}{Module ID}
}
\description{
Module UI for data upload
}
\keyword{internal}
