% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-data-wrangling.R
\name{module_ui_wrangle_data}
\alias{module_ui_wrangle_data}
\title{Module UI for data upload}
\usage{
module_ui_wrangle_data(id)
}
\arguments{
\item{id}{Module ID}
}
\description{
Module UI for data upload
}
\keyword{internal}
