% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_clusters}
\alias{parse_clusters}
\title{Extract results from SaTScan-text-based output}
\usage{
parse_clusters(file)
}
\arguments{
\item{file}{SaTScan-text-based output result given as "main" to be parsed}
}
\description{
Extract results from SaTScan-text-based output
}
\keyword{internal}
