% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runApp.R
\name{ww_run_app}
\alias{ww_run_app}
\title{Initialise built-in Shiny application}
\usage{
ww_run_app(package = "wowi")
}
\arguments{
\item{package}{package name ("wowi").}
}
\value{
Called for its side effect of launching the \code{wowi} Shiny
application in the user's default web browser. Returns \code{NULL}
invisibly when the app session ends.
}
\description{
Initialise built-in Shiny application
}
\examples{
if(interactive()) ww_run_app()

}
