% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{anthro}
\alias{anthro}
\title{Sample data set of district-level SMART surveys}
\format{
A tibble of 2,934 rows and 17 columns.
}
\source{
anonymous
}
\usage{
anthro
}
\description{
\code{anthro} is a SMART survey-generated data conducted in nine districts in Uganda.
Some districts
}
\examples{
anthro

}
\keyword{datasets}
