% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wowi-package.R
\docType{package}
\name{wowi-package}
\alias{wowi}
\alias{wowi-package}
\title{wowi: Detect Spatial Clusters of High Acute Malnutrition Rates}
\description{
Utilities for detecting statistically significant spatial clusters of high acute malnutrition rates using SaTScan's Bernoulli spatial-scan model.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/nutspatial/wowi}
  \item Report bugs at \url{https://github.com/nutspatial/wowi/issues}
}

}
\author{
\strong{Maintainer}: Tomás Zaba \email{tomas.zaba@outlook.com} (\href{https://orcid.org/0000-0002-7079-3574}{ORCID}) [copyright holder]

}
\keyword{internal}
