% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wowi-package.R
\docType{package}
\name{wowi-package}
\alias{wowi}
\alias{wowi-package}
\title{wowi: Detect Spatial Clusters of High Rates of Acute Malnutrition}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Utilities for detecting statistically significant spatial clusters of high acute malnutrition rates using a Bernoulli spatial scan statistic, implemented via the SaTScan software (\url{https://www.satscan.org/}).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tiwowi/wowi}
  \item \url{https://tiwowi.github.io/wowi/}
  \item Report bugs at \url{https://github.com/tiwowi/wowi/issues}
}

}
\author{
\strong{Maintainer}: Tomás Zaba \email{tomas.zaba@outlook.com} (\href{https://orcid.org/0000-0002-7079-3574}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Tomás Zaba \email{tomas.zaba@outlook.com} (\href{https://orcid.org/0000-0002-7079-3574}{ORCID}) [copyright holder]
}

}
\keyword{internal}
