% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/satscan-configuration.R
\name{ww_configure_satscan}
\alias{ww_configure_satscan}
\title{Configure SaTScan for a Bernoulli purely spatial scan}
\usage{
ww_configure_satscan(
  filename = character(),
  params_destfile = character(),
  satscan_version = character(),
  .scan_for = c("high-rates", "high-low-rates")
)
}
\arguments{
\item{filename}{A quoted name identifying the analysis area.}

\item{params_destfile}{A quoted name of the folder or directory where the
parameters file (produced by this function) should be saved. This can be
the same directory as that specified in \code{\link[=ww_wrangle_data]{ww_wrangle_data()}}.}

\item{satscan_version}{A quoted value representing the version of SaTScan
installed on the user's computer. Internally, this value is checked against
the latest available version. If it is older, a warning is issued with a link
to SaTScan’s website. Although the analysis is not interrupted, it is
recommended to use the latest version.}

\item{.scan_for}{A string indicating the type of clusters to scan for.
To scan for high-rate clusters only, set \code{.scan_for = "high-rates"}.
To scan for both high and low rates, set \code{.scan_for = "high-low-rates"}.}
}
\description{
Defines the analysis parameters required by SaTScan's GUI to conduct a
Bernoulli-based purely spatial scan, either to detect clusters of high rates
of acute malnutrition or both high and low rates.

User input is limited to specifying the analysis area filename, the destination
directory for the parameters file, the SaTScan version in use, and the type of
clusters to be detected. All other parameters are pre-defined by this function.
}
\details{
For more information on Bernoulli purely spatial scans, refer to the
SaTScan technical documentation: \url{https://www.satscan.org/techdoc.html}
}
\examples{
## Given a temporary directory ----
tmp <- withr::local_tempdir()
directory <- file.path(tmp, "input-files")

## Wrangle data with `{mwana}` ----
x <- df |>
  mwana::mw_wrangle_wfhz(
    sex = sex,
    .recode_sex = TRUE,
    weight = weight,
    height = height
  ) |>
  mwana::define_wasting(
    zscores = wfhz,
    .by = "zscores",
    edema = edema
  )

## Apply the function ----
ww_wrangle_data(
  .data = x,
  filename = "Locality",
  destfile = directory,
  .gam_based = "wfhz"
)

#### Configure SaTScan ----
do.call(
  what = ww_configure_satscan,
  args = list(
    filename = "Locality",
    params_destfile = directory,
    satscan_version = "10.3.2",
    .scan_for = "high-low-rates"
  )
)

## Show file's content ----
file.show(file.path(tmp, "input-files/Locality.prm"))

}
