% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runApp.R
\name{ww_run_app}
\alias{ww_run_app}
\title{Initialise built-in Shiny application}
\usage{
ww_run_app(pkg = "wowi")
}
\description{
Initialise built-in Shiny application
}
\examples{
if(interactive()) ww_run_app()

}
